Replaces godot networking with Riptide networking in the connection and loading of the level part

+++ Multiplayer/PlayerState.cs
@@ -4,6 +4,6 @@ using System;
 public partial class PlayerState : GodotObject
 {
 	public string Name;
-	public long Id;
+	public ushort Id;
 	public bool IsLoading;
 }
